Revert "Thread-Safe Implementation of Slab (#68)"
This reverts commit af9241e0386db7778cde990ee9e111590938984e.

We are reverting the additional multithreading support from libusual
master for now.  They will be resubmitted to the pgbouncer repository
after the submodules are removed there.
