+++ test/test_common.c
@@ -35,7 +35,6 @@ struct testgroup_t groups[] = {
 	{ "regex/", regex_tests },
 	{ "shlist/", shlist_tests },
 	{ "socket/", socket_tests },
-	{ "slab_ts/", slab_ts_tests },
 	{ "string/", string_tests },
 	{ "strpool/", strpool_tests },
 	{ "talloc/", talloc_tests },
