+++ test/test_common.h
@@ -51,7 +51,6 @@ extern struct testcase_t psrandom_tests[];
 extern struct testcase_t pthread_tests[];
 extern struct testcase_t regex_tests[];
 extern struct testcase_t shlist_tests[];
-extern struct testcase_t slab_ts_tests[];
 extern struct testcase_t socket_tests[];
 extern struct testcase_t string_tests[];
 extern struct testcase_t strpool_tests[];
